style: refactor functions to where they should be
