+++ headers/dump_section_header.h
@@ -29,11 +29,8 @@ typedef struct section_header {
 void navigate_fd_to_section_header(Elf_header header, FILE *fd);
 void navigate_fd_to_section_index(Elf_header header, FILE *fd, int index);
 Section_header grab_sect_header(Elf_header header, Args args, int index);
-uint64_t read_nbytes_better(Elf_header header, FILE *fd, int bytes, bool variable);
-void DEBUG_DUMP_NBYTES(int offset, int n, Args args);
 Section_header *grab_all_section_headers(Elf_header header, Args args);
 void dump_section_headers(Section_header *headers, Elf_header elf_header, Args args);
 void print_and_format_section_header(Section_header shname, Section_header h, Elf_header elf_header, int i, Args args);
-void read_stream_until_null(FILE *fd);
 
 #endif
