+++ src/dump_elf_header.c
@@ -15,75 +15,6 @@ const char *object_type_names[5] = {
     "Unknown", "Relocatable", "Executable", "Shared Object", "Core"
 };
 
-/*
-// reads 32/64 bits depending of class in elf_header
-uint64_t read_varaible_entries(FILE *fd, Elf_header header) {
-    uint64_t ret = 0;
-    if (header.class == 1) {
-        ret = getc(fd);
-    } else if (header.class == 2) {
-        ret = getc(fd);
-        ret <<= 32;
-        ret = getc(fd);
-    } else {
-        fatal_error("ERROR: elf header class is neither 1 nor 2");
-    }
-
-    return ret;
-}
-*/
-
-// IM SO UGLY 😭 sigh... 💔
-// is goto aura?
-// depreciated by read_nbytes_better lmao
-uint64_t read_nbytes(FILE *fd, Elf_header *header, int byte_count, bool variable) {
-    uint64_t ret = 0;
-
-    if (variable) {
-        if (header->ei_class == 1) {
-            byte_count = 4;
-        } else {
-            byte_count = 8;
-        }
-    }
-
-    // little endian
-    if (header->ei_data == 1) {
-        ret |= getc(fd);
-        ret |= ((uint64_t) getc(fd) << (8));
-        if (byte_count == 2) { goto READ_NBYTES__END; }
-        
-        ret |= ((uint64_t) getc(fd) << (8 * 2));
-        ret |= ((uint64_t) getc(fd) << (8 * 3));
-        if (byte_count == 4) { goto READ_NBYTES__END; }
-        
-        ret |= ((uint64_t) getc(fd) << (8 * 4));
-        ret |= ((uint64_t) getc(fd) << (8 * 5));
-        ret |= ((uint64_t) getc(fd) << (8 * 6));
-        ret |= ((uint64_t) getc(fd) << (8 * 7));
-        if (byte_count == 8) { goto READ_NBYTES__END; }
-    } else if (header->ei_data == 2) {
-        ret |= getc(fd);
-        ret <<= 8; ret |= getc(fd);
-        if (byte_count == 4) { goto READ_NBYTES__END; }
-
-        ret <<= 8; ret |= getc(fd);
-        ret <<= 8; ret |= getc(fd);
-        if (byte_count == 4) { goto READ_NBYTES__END; }
-
-        ret <<= 8; ret |= getc(fd);
-        ret <<= 8; ret |= getc(fd);
-        ret <<= 8; ret |= getc(fd);
-        ret <<= 8; ret |= getc(fd);
-        if (byte_count == 8) { goto READ_NBYTES__END; }
-    } else {
-        fatal_error("ERROR: expected endianess to be 1 or 2.");
-    }
-
-READ_NBYTES__END:
-    return ret;
-}
-
 
 void check_and_set_magic(FILE *fd, Elf_header *header) {
     header->ei_magic[0] = getc(fd);
